+++ src/renderer/src/utils/index.ts
@@ -4,7 +4,6 @@ export { folderTreeChildren, folderTreePaddingLeft } from './folder-tree';
 export { formatFileSize } from './format-file-size';
 export { formatResolution } from './format-resolution';
 export { formatCreatedAt, formatMetadataValue } from './format-metadata-value';
-export { getGridThumbnailSize } from './grid-thumbnail';
 export { getThumbnailColumnCount } from './thumbnail-columns';
 export { buildMetadataRows, groupMetadataRows, type MetadataGroup } from './group-metadata';
 export { getPhotoOverview, type PhotoOverviewItem } from './photo-overview';